requirements.txt
numpy==2.0.0
certifi==2024.7.4
aiohttp[speedups]==3.9.5
hikari[speedups]==2.0.0.dev126
hikari-lightbulb==2.3.5
hikari-miru==4.1.1
lavalink-rs==0.12.0
PyYAML==6.0
requests
pillow==10.3.0
google-api-python-client==2.136.0
schedule==1.2.1
nekos.py==1.1.0
selenium==4.20.0
webdriver-manager==4.0.1
rule34Py==1.4.11
pytz==2024.1
captcha==0.5.0
aiofiles==23.2.1
matplotlib==3.9.0
segno==1.6.1
qrcode-artistic==3.0.2
psutil==6.0.0
Flask==3.0.3